requirements.txt
altgraph==0.17.4
asgiref==3.7.2
bidict==0.23.1
blinker==1.7.0
branca==0.7.1
build==1.2.1
CacheControl==0.14.0
certifi==2023.7.22
charset-normalizer==3.2.0
chess==1.10.0
cleo==2.1.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
crashtest==0.4.1
cycler==0.12.1
distlib==0.3.8
Django==5.0.3
dnspython==2.6.1
dulwich==0.21.7
eventlet==0.36.1
fastjsonschema==2.19.1
filelock==3.14.0
Flask==3.0.1
Flask-SocketIO==5.3.6
folium==0.16.0
fonttools==4.47.0
greenlet==3.0.3
h11==0.14.0
idna==3.4
igraph==0.10.8
importlib_metadata==7.1.0
iniconfig==2.0.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.4.0
Jinja2==3.1.3
keyring==24.3.1
kiwisolver==1.4.5
MarkupSafe==2.1.4
matplotlib==3.8.2
more-itertools==10.2.0
msgpack==1.0.8
networkx==3.1
numpy==1.26.2
packaging==23.2
pefile==2023.2.7
pexpect==4.9.0
Pillow==10.1.0
pkginfo==1.10.0
platformdirs==4.2.1
plotly==5.18.0
pluggy==1.3.0
poetry-core==1.9.0
poetry-plugin-export==1.7.1
polyline==2.0.2
ptyprocess==0.7.0
pyinstaller==6.7.0
pyinstaller-hooks-contrib==2024.6
pyparsing==3.1.1
pyproject_hooks==1.1.0
pytest==7.4.3
python-chess==1.999
python-dateutil==2.8.2
python-dotenv==1.0.1
python-engineio==4.9.1
python-louvain==0.16
python-socketio==5.11.3
pywin32-ctypes==0.2.2
rapidfuzz==3.9.0
requests==2.31.0
requests-toolbelt==1.0.0
scipy==1.13.0
shellingham==1.5.4
simple-websocket==1.0.0
six==1.16.0
sqlparse==0.4.4
tenacity==8.2.3
texttable==1.6.7
tomlkit==0.12.4
trove-classifiers==2024.4.10
tzdata==2024.1
urllib3==2.0.4
virtualenv==20.26.1
websocket-client==1.8.0
Werkzeug==3.0.1
wsproto==1.2.0
xyzservices==2023.10.1
zipp==3.18.1
